fixed (?) extractCredPatch
Nessus plugin 66334 (Patch Report) is also flagged when a remote version enum (nmap, http headers, banners) happens. So this is not good. At the same time, there is no single plugin to filter for.

So, in order to find out which hosts got cred patched I am now looking for:
141118,     # Target Credential Status by Authentication Protocol - Valid Credentials Provided
97993,      # OS Identification and Installed Software Enumeration over SSH v2 (Using New SSH Library)
152742      # Unix Software Discovery Commands Available

!!There might be more to look for!!